+++ flask_ai/resources/framework/chatbot.py
@@ -4,14 +4,14 @@ import os
 import time
 
 import openai
-
 from dotenv import load_dotenv
-from colors import pr_bot_response
-from debug_utils import debug_steps, debug, debug_attribute
-from constants import davinci, HUMAN, BOT, SEPARATORS, INITIAL_PROMPT, INITIAL_RESPONSE, COST, fields_dict
-from utils import get_db_connection, replace_quotes, write_to_db, write_logs_to_csv
-from app.utils import chatbot_logic, get_props_from_message
-from app.constants import RESPONSE_FOR_INVALID_QUERY, UNABLE_TO_FIND_PRODUCTS_IN_DB, CHATBOT_NAME
+
+from .colors import pr_bot_response, pr_red
+from .debug_utils import debug_steps, debug, debug_attribute
+from .constants import davinci, HUMAN, BOT, INITIAL_PROMPT, INITIAL_RESPONSE, COST, fields_dict
+from .utils import get_db_connection, replace_quotes, write_to_db, write_logs_to_csv, get_last_n_message_log
+from .app.utils import chatbot_logic, get_props_from_message
+from .app.constants import RESPONSE_FOR_INVALID_QUERY, UNABLE_TO_FIND_PRODUCTS_IN_DB, CHATBOT_NAME
 
 # Insert your API key
 load_dotenv()
@@ -27,7 +27,7 @@ conn, cur = get_db_connection()
 
 def get_answer_from_gpt(row, prompt, level):
     # Multi shot learning
-    TOKENS, TEMPERATURE, MODEL, STOP = 300, 0, davinci, [HUMAN, BOT]
+    TOKENS, TEMPERATURE, MODEL, STOP = 200, 0, davinci, [HUMAN, BOT]
     response = openai.Completion.create(
         model=MODEL,
         prompt=prompt,
@@ -43,7 +43,20 @@ def get_answer_from_gpt(row, prompt, level):
     return response_text, response_token
 
 
-def chatbot(user_input, message_log, time_stamp, discord_id="", db=True):
+def get_chat_response(user_input, message_log, time_stamp, html_response, discord_id, db=True):
+    updated_message_log, row_id = "", ""
+    if len(user_input) > 300:
+        response = "Please type a message that is less than 300 characters."
+        pr_red(response)
+    else:
+        response, updated_message_log, row_id = chatbot(
+            user_input, message_log, time_stamp, html_response, discord_id, db)
+    return response, updated_message_log, row_id
+
+
+def chatbot(user_input, message_log, time_stamp, html_response, discord_id, db):
+    # Store only last 2 conversation and prompt conversation
+    message_log = get_last_n_message_log(message_log, 2)
     MODE = 'w'
     fields = ["user_input", "bot_response", "level1", "level2", "level3", "level4",
               INITIAL_PROMPT, INITIAL_RESPONSE, COST]
@@ -56,9 +69,7 @@ def chatbot(user_input, message_log, time_stamp, discord_id="", db=True):
     if os.path.exists(DEBUG_CSV):
         MODE = 'a'
 
-    # Append user_input
-    prompt = f"{message_log}{HUMAN}{user_input}\n{BOT}"
-    prompt = prompt.replace(SEPARATORS, '')
+    prompt = "\n".join(message_log) + f"{HUMAN}{user_input}{BOT}"
 
     debug_steps(row, f"Prompt - {prompt}", level=INITIAL_PROMPT)
 
@@ -89,7 +100,7 @@ def chatbot(user_input, message_log, time_stamp, discord_id="", db=True):
         valid_query = False
     else:
         bot_response, raw_response, tokens = chatbot_logic(
-            row, user_input, raw_gpt_response)
+            row, user_input, raw_gpt_response, html_response)
         query_to_tokens = tokens
 
     if ((not bot_response or len(bot_response) < 10) and bot_response != RESPONSE_FOR_INVALID_QUERY):
@@ -105,8 +116,7 @@ def chatbot(user_input, message_log, time_stamp, discord_id="", db=True):
 
     record = replace_quotes([user_input, bot_response, prompt, raw_gpt_response, row[fields_dict[1]], row[fields_dict[2]],
                             row[fields_dict[3]], row[fields_dict[4]], response_accepted, response_time, discord_id, cost, time_stamp])
-
-    write_to_db(db, record)
+    row_id = write_to_db(db, record)
 
     debug(f"Response time in seconds - {response_time}")
 
@@ -114,8 +124,8 @@ def chatbot(user_input, message_log, time_stamp, discord_id="", db=True):
 
     # Add the chatbot's response to the conversation history
     if raw_gpt_response not in UNABLE_TO_FIND_PRODUCTS_IN_DB and valid_query and raw_gpt_response not in RESPONSE_FOR_INVALID_QUERY:
-        message_log += f"{HUMAN}{user_input}\n{BOT}{raw_response}{SEPARATORS}"
+        message_log.append(f"{HUMAN}{user_input}\n{BOT}{raw_response}")
 
     pr_bot_response(bot_response)
     print(f"Total cost - {cost}")
-    return bot_response, message_log
+    return bot_response, message_log, row_id